added the illustrative triangle dataset
it illustrates the inability of GCNs to aggregate scores assigned to connected substructures with related labels.
this becomes apparent when connecting nodes that should be part of a single constituent.
=> static neighborhood circles underperform here.

next step: evaluate a basic learned spatial edge filter to enhance the local spectral filtering that is currently performed to obtain irregular constituent boundaries (this can be interpreted as learned LTA decompositions).